scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
selenium==4.15.0
fake-useragent==1.4.0
retrying==1.3.4
pandas==2.1.0